cleaning

+++ src/API.DeskBookService.Data/Repository/DeskRepository.cs
@@ -9,10 +9,17 @@ namespace API.DeskBookService.Data.Repository
     public class DeskRepository : IDeskRepository
     {
         private IMongoCollection<Desk> _desks;
+        private IMongoCollection<DeskBooking> _bookings;
 
         public DeskRepository(IDeskBookerDataContext deskBookerContext)
         {
             _desks = deskBookerContext.Desks;
+            _bookings = deskBookerContext.DesksBooking;
+        }
+        public async Task<Desk> Save(Desk desk)
+        {
+            await _desks.InsertOneAsync(desk);
+            return desk;
         }
 
         public async Task<Desk> Get(string id)
@@ -27,14 +34,12 @@ namespace API.DeskBookService.Data.Repository
 
         public async Task<bool> Remove(string id)
         {
+            var booking = await _bookings.Find(b => b.DeskId == id).AnyAsync();
+            if (booking)
+                return false;
+
             var result = await _desks.DeleteOneAsync(desk => desk.Id == id);
             return result.IsAcknowledged && result.DeletedCount > 0;
-        }
-
-        public async Task<Desk> Save(Desk desk)
-        {
-            await _desks.InsertOneAsync(desk);
-            return desk;
         }
 
         public async Task<bool> Update(string id, Desk deskIn)
